Refactor saveToFile boolean in Hologram constructor
Essentially made it do what it was supposed to do from the start.

If saveToFile is false, the FileConfig will be null and not initialized.

The reason why this was a problem in the first place was because an inline if statement was being used. Assuming saveToFile was set to faslse, due to java's type checking, this was causing a FileConfig object to be initialized without being used, and as a consequence an empty file would be generated. This is now fixed in this commit.